audio_to_sign/requirements.txt
annotated-types==0.6.0
av==9.1.1
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
filelock==3.13.1
fsspec==2023.12.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
preshed==3.0.9
proglog==0.1.10
PyAudio==0.2.14
pydantic==2.5.2
pydantic-core==2.14.5
pydub==0.25.1
pygame==2.5.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.1
tkvideoplayer==2.5
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing-extensions==4.8.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
